refactored flatten defects to defect service
