+++ src/services/DefectService.ts
@@ -2,6 +2,10 @@ import { Service } from 'typedi';
 import moment from 'moment';
 import { ICustomDefect } from '../models/ICustomDefect';
 import { IVA_30 } from '../models/Enums';
+import { IFlatDefect } from '../models/IFlatDefect';
+import { IItem } from '../models/IItem';
+import { IDefectParent } from '../models/IDefectParent';
+import { IDefectChild } from '../models/IDefectChild';
 
 @Service()
 export class DefectService {
@@ -25,4 +29,62 @@ export class DefectService {
     };
     return (customDefects && customDefects.length > 0) ? customDefects : [defaultCustomDefect];
   };
+
+  /**
+   * Returns a flattened array of every deficiency that only includes the key/value pairs required for certificate generation
+   * @param defects - the array of defects from the api
+   */
+  public flattenDefectsFromApi(defects: IDefectParent[]): IFlatDefect[] {
+    const flatDefects: IFlatDefect[] = [];
+    try {
+      // go through each defect in un-flattened array
+      defects.forEach((defect: IDefectParent) => {
+        const {
+          imNumber, imDescription, imDescriptionWelsh, items,
+        } = defect;
+        if (defect.items !== undefined && defect.items.length !== 0) {
+          // go through each item of defect
+          items.forEach((item: IItem) => {
+            const {
+              itemNumber,
+              itemDescription,
+              itemDescriptionWelsh,
+              deficiencies,
+            } = item;
+            if (
+              item.deficiencies !== undefined
+              && item.deficiencies.length !== 0
+            ) {
+              // go through each deficiency and push to flatDefects array
+              deficiencies.forEach((deficiency: IDefectChild) => {
+                const {
+                  ref,
+                  deficiencyText,
+                  deficiencyTextWelsh,
+                  forVehicleType,
+                } = deficiency;
+                const lowLevelDeficiency: IFlatDefect = {
+                  imNumber,
+                  imDescription,
+                  imDescriptionWelsh,
+                  itemNumber,
+                  itemDescription,
+                  itemDescriptionWelsh,
+                  ref,
+                  deficiencyText,
+                  deficiencyTextWelsh,
+                  forVehicleType,
+                };
+                flatDefects.push(lowLevelDeficiency);
+              });
+            }
+          });
+        }
+      });
+    } catch (e) {
+      // eslint-disable-next-line
+      console.error(`Error flattening defects: ${e}`);
+    }
+    return flatDefects;
+  }
 }
